Rename some components according to review
