+++ src/plugins/components/PluginDetailsChannelsCard/PluginDetailsChannelsCard.tsx
@@ -1,93 +1,21 @@
-import {
-  CardContent,
-  Divider,
-  makeStyles,
-  Typography
-} from "@material-ui/core";
-import Skeleton from "@saleor/components/Skeleton";
-import { renderCollectionWithDividers } from "@saleor/misc";
-import { Plugin_plugin } from "@saleor/plugins/types/Plugin";
-import { isPluginGlobal } from "@saleor/plugins/views/utils";
+import { Card } from "@material-ui/core";
+import CardTitle from "@saleor/components/CardTitle";
+import { sectionNames } from "@saleor/intl";
 import React from "react";
-import { FormattedMessage } from "react-intl";
+import { useIntl } from "react-intl";
 
-const useStyles = makeStyles(
-  theme => ({
-    itemContainer: {
-      position: "relative",
-      cursor: "pointer"
-    },
-    itemActiveIndicator: {
-      position: "absolute",
-      left: 0,
-      backgroundColor: theme.palette.primary.main,
-      width: 2,
-      height: "100%"
-    }
-  }),
-  { name: "PluginDetailsChannelsCard" }
-);
+import PluginDetailsChannelsCardContent, {
+  PluginDetailsChannelsCardProps
+} from "./PluginDetailsChannelsCardContent";
 
-interface PluginDetailsChannelsCardProps {
-  setSelectedChannelId: (channelId: string) => void;
-  selectedChannelId: string;
-  plugin: Plugin_plugin;
-}
-
-import { pluginDetailsChannelsCardMessages as messages } from "./messages";
-import PluginDetailsChannelsCardContainer from "./PluginDetailsChannelsCardContainer";
-
-const PluginDetailsChannelsCard: React.FC<PluginDetailsChannelsCardProps> = ({
-  plugin,
-  selectedChannelId,
-  setSelectedChannelId
-}) => {
-  const classes = useStyles({});
-
-  if (!plugin) {
-    return (
-      <PluginDetailsChannelsCardContainer>
-        <CardContent>
-          <Skeleton />
-        </CardContent>
-      </PluginDetailsChannelsCardContainer>
-    );
-  }
-
-  if (isPluginGlobal(plugin.globalConfiguration)) {
-    return (
-      <PluginDetailsChannelsCardContainer>
-        <CardContent>
-          <FormattedMessage {...messages.noChannelsSubtitle} />
-        </CardContent>
-      </PluginDetailsChannelsCardContainer>
-    );
-  }
-
-  const isChannelSelected = (channelId: string) =>
-    selectedChannelId === channelId;
+const PluginDetailsChannelsCard: React.FC<PluginDetailsChannelsCardProps> = props => {
+  const intl = useIntl();
 
   return (
-    <PluginDetailsChannelsCardContainer>
-      {renderCollectionWithDividers({
-        collection: plugin.channelConfigurations,
-        renderDivider: () => <Divider />,
-        renderItem: ({ channel }) => (
-          <div
-            className={classes.itemContainer}
-            key={channel.id}
-            onClick={() => setSelectedChannelId(channel.id)}
-          >
-            {isChannelSelected(channel.id) && (
-              <div className={classes.itemActiveIndicator}></div>
-            )}
-            <CardContent>
-              <Typography>{channel.name}</Typography>
-            </CardContent>
-          </div>
-        )
-      })}
-    </PluginDetailsChannelsCardContainer>
+    <Card>
+      <CardTitle title={intl.formatMessage(sectionNames.channels)} />
+      <PluginDetailsChannelsCardContent {...props} />
+    </Card>
   );
 };
